# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dashboard Platform API
  version: 1.0.0
  description: >-
    The official platform contract for managing inbox sources, tasks,
    immutable artifacts, approvals, and audit history around the dashboard.
servers:
  - url: http://localhost:8000/api/v1
    description: Local service
  - url: http://127.0.0.1:8000/api/v1
    description: Local service by IP
security:
  - bearerAuth: []
...
paths:
...
  /tasks/:
    get:
      tags: [Tasks]
      summary: List tasks
      parameters:
        - name: status
          in: query
          schema: {$ref: "#/components/schemas/TaskStatus"}
        - name: assigned_agent
          in: query
          schema: {type: string}
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Tasks]
      summary: Create a task
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: {$ref: "#/components/schemas/Task"}
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
...
  /artifacts/:
    get:
      tags: [Artifacts]
      summary: List artifact versions
      parameters:
        - name: task_id
          in: query
          schema: {type: string, format: uuid}
        - name: series_id
          in: query
          schema: {type: string, format: uuid}
        - name: kind
          in: query
          schema: {$ref: "#/components/schemas/ArtifactKind"}
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Artifact list without content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactListResponse"
    post:
      tags: [Artifacts]
      summary: Create an immutable artifact revision
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtifactCreate"
      responses:
        "201":
          description: Created artifact version
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: {$ref: "#/components/schemas/ArtifactWithContent"}
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
...
  /approvals/:
    get:
      tags: [Approvals]
      summary: List approval requests
      parameters:
        - name: task_id
          in: query
          schema: {type: string, format: uuid}
        - name: status
          in: query
          schema: {$ref: "#/components/schemas/ApprovalStatus"}
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Approval list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalListResponse"
    post:
      tags: [Approvals]
      summary: Request approval of a specific artifact
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalCreate"
      responses:
        "201":
          description: Approval requested
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: {$ref: "#/components/schemas/Approval"}
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: dpt
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
...
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 50}
    Offset:
      name: offset
      in: query
      schema: {type: integer, minimum: 0, default: 0}
  responses:
    Unauthorized:
      description: Missing, invalid, expired or revoked token
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Forbidden:
      description: Insufficient scope
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
...
    Conflict:
      description: State, idempotency or immutable artifact conflict
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    PreconditionRequired:
      description: Idempotency-Key is required
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            details: {}
    Pagination:
      type: object
      required: [count, limit, offset, next_offset]
      properties:
        count: {type: integer, minimum: 0}
        limit: {type: integer}
        offset: {type: integer}
        next_offset: {type: [integer, "null"]}
...
    TaskStatus:
      type: string
      enum:
        - collected
        - analyzing
        - draft
        - needs_review
        - approved
        - rejected
        - revision_requested
        - queued
        - executing
        - completed
        - failed
    TaskCreate:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title: {type: string, maxLength: 250}
        description: {type: string}
        priority: {type: string, enum: [low, normal, high, urgent], default: normal}
        inbox_item_id: {type: string, format: uuid}
        assigned_agents:
          type: array
          items: {type: string}
        metadata: {type: object, additionalProperties: true}
        due_at: {type: string, format: date-time}
...
    Task:
      allOf:
        - $ref: "#/components/schemas/TaskCreate"
        - type: object
          required: [id, status, created_by, created_at, updated_at]
          properties:
            id: {type: string, format: uuid}
            status: {$ref: "#/components/schemas/TaskStatus"}
            status_label: {type: string}
            priority_label: {type: string}
            created_by: {type: string}
            completed_at: {type: [string, "null"], format: date-time}
            created_at: {type: string, format: date-time}
            updated_at: {type: string, format: date-time}
    TaskListResponse:
      type: object
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/Task"}
        pagination: {$ref: "#/components/schemas/Pagination"}
    ArtifactKind:
      type: string
      enum: [source, analysis, draft, answer, report, result, other]
    ArtifactCreate:
      type: object
      additionalProperties: false
      required: [task_id, kind, title, content]
      properties:
        task_id: {type: string, format: uuid}
        series_id: {type: string, format: uuid}
        kind: {$ref: "#/components/schemas/ArtifactKind"}
        title: {type: string, maxLength: 250}
        content: {type: string}
        mime_type: {type: string, default: text/markdown}
        metadata: {type: object, additionalProperties: true}
    Artifact:
      type: object
      required:
        - id
        - series_id
        - version
        - task_id
        - kind
        - title
        - storage_uri
        - sha256
        - size_bytes
        - created_by
        - created_at
      properties:
        id: {type: string, format: uuid}
        series_id: {type: string, format: uuid}
        version: {type: integer, minimum: 1}
        task_id: {type: string, format: uuid}
        kind: {$ref: "#/components/schemas/ArtifactKind"}
        kind_label: {type: string}
        title: {type: string}
        mime_type: {type: string}
        storage_uri: {type: string}
        sha256: {type: string, pattern: "^[0-9a-f]{64}$"}
        size_bytes: {type: integer, minimum: 0}
        created_by: {type: string}
        metadata: {type: object, additionalProperties: true}
        created_at: {type: string, format: date-time}
    ArtifactWithContent:
      allOf:
        - $ref: "#/components/schemas/Artifact"
        - type: object
          required: [content]
          properties:
            content: {type: string}
    ArtifactListResponse:
      type: object
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/Artifact"}
        pagination: {$ref: "#/components/schemas/Pagination"}
    ApprovalStatus:
      type: string
      enum: [pending, approved, rejected, revision_requested]
    ApprovalCreate:
      type: object
      additionalProperties: false
      required: [task_id, artifact_id]
      properties:
        task_id: {type: string, format: uuid}
        artifact_id: {type: string, format: uuid}
        note: {type: string}
...
    Approval:
      type: object
      required:
        - id
        - task_id
        - artifact_id
        - target_sha256
        - status
        - requested_by
        - requested_at
      properties:
        id: {type: string, format: uuid}
        task_id: {type: string, format: uuid}
        artifact_id: {type: string, format: uuid}
        target_sha256: {type: string, pattern: "^[0-9a-f]{64}$"}
        status: {$ref: "#/components/schemas/ApprovalStatus"}
        status_label: {type: string}
        request_note: {type: string}
        decision_note: {type: string}
        requested_by: {type: string}
        decided_by: {type: [string, "null"]}
        requested_at: {type: string, format: date-time}
        decided_at: {type: [string, "null"], format: date-time}
    ApprovalListResponse:
      type: object
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/Approval"}
        pagination: {$ref: "#/components/schemas/Pagination"}
...